docker_scaping_berita/requirements.txt
beautifulsoup4==4.12.2
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
selenium==4.14.0
webdriver-manager==4.0.1 